split apart giving saga
