+++ imports/store/give/saga/chargeTransaction.js
@@ -1,0 +1,119 @@
+
+import "regenerator-runtime/runtime";
+import { take, put, cps, select } from "redux-saga/effects";
+import gql from "graphql-tag";
+
+import { schedule, charge } from "../../../methods/give/browser";
+import { GraphQL } from "../../../graphql";
+
+import types from "./../types";
+import actions from "../actions";
+
+import formatPersonDetails from "./formatPersonDetails";
+import validate from "./validate";
+import CREATE_ORDER_MUTATION from "./createOrderMutation";
+import COMPLETE_ORDER_MUTATION from "./completeOrderMutation";
+import submitPersonDetails from "./submitPersonDetails";
+
+
+// handle the transactions
+export default function* chargeTransaction({ state }) {
+  if (state !== "submit") return;
+
+  let { give } = yield select();
+  const name = give.data.payment.name;
+  let action = charge;
+  let error = false;
+  let id;
+  let saved = false;
+
+  // set loading state
+  yield put(actions.loading());
+  // personal info is ready to be submitted
+  const formattedData = formatPersonDetails(give);
+
+  // if you have a saved account, NMI lets you "order" a schedule
+  // instead of order + charge
+  if (formattedData.savedAccount && Object.keys(give.schedules).length) {
+    saved = true;
+  } else {
+    let store = yield select();
+    give = store.give;
+
+    if (formattedData.savedAccount) {
+      // set people data and store transaction id
+      yield* submitPersonDetails(give, true);
+    }
+
+    store = yield select();
+    give = store.give;
+
+    // wait until we have the transaction url
+    if (!give.url) {
+      const { url } = yield take(types.SET_TRANSACTION_DETAILS);
+      give.url = url;
+    }
+  }
+
+  // get the token and name of the saved account
+  const token = give.url.split("/").pop();
+
+  if (Object.keys(give.schedules).length) {
+    // if there is not a saved account, charge the order
+    if (!formattedData.savedAccount) {
+      action = schedule;
+
+      if (give.data.payment.type === "cc") {
+        // saved accounts don't validate the payment by default
+        // so we make 3 blocking requests to validate the card :(
+        const { validationError } = yield* validate();
+
+        if (validationError) {
+          error = validationError;
+        }
+      }
+    }
+  }
+
+
+  if (give.scheduleToRecover && Object.keys(give.schedules).length) {
+    id = give.scheduleToRecover;
+  }
+
+  if (give.userId) id = give.userId; // used in native => web payments
+
+  if (!error) {
+    // submit transaction
+    try {
+      if (saved) {
+        // XXX update data if returned
+        yield GraphQL.mutate({
+          mutation: COMPLETE_ORDER_MUTATION,
+          variables: { token, name, id },
+        })
+          .then(x => { console.log(x); });
+      } else {
+        yield GraphQL.mutate({
+          mutation: CREATE_ORDER_MUTATION,
+          variables: { data: JSON.stringify(formattedData), instant: true, id },
+        });
+      }
+    } catch (e) { error = e; }
+  }
+  // set error states
+  if (error) {
+    yield put(actions.error({ transaction: error }));
+
+    // remove loading state
+    yield put(actions.setState("error"));
+  } else {
+    // remove loading state
+    yield put(actions.setState("success"));
+
+    // if we activated an inactive schedule, remove it
+    if (give.scheduleToRecover && give.recoverableSchedules[give.scheduleToRecover]) {
+      yield put(actions.deleteSchedule(give.scheduleToRecover));
+      yield put(actions.deleteRecoverableSchedules(give.scheduleToRecover));
+    }
+  }
+}
