+++ imports/store/give/saga/completeOrderMutation.js
@@ -1,0 +1,11 @@
+import gql from "graphql-tag";
+
+export default gql`
+  mutation completeOrder($token: String!, $id: ID, $name: String) {
+    response: completeOrder(token: $token, id: $id, name: $name) {
+      error
+      success
+      code
+    }
+  }
+`;
