+++ imports/store/give/saga/createOrderMutation.js
@@ -1,0 +1,13 @@
+import gql from "graphql-tag";
+
+export default gql`
+  mutation order($data: String!, $id: ID, $instant: Boolean) {
+    response: createOrder(data: $data, id: $id, instant: $instant) {
+      url
+      error
+      success
+      code
+      transactionId
+    }
+  }
+`;
