+++ imports/store/give/saga/index.js
@@ -1,11 +1,8 @@
-/* eslint-disable */
 import "regenerator-runtime/runtime";
 import { Meteor } from "meteor/meteor";
-import React from "react";
-import ReactDOM from "react-dom";
-import Moment from "moment";
+import moment from "moment";
 import { takeLatest, takeEvery } from "redux-saga";
-import { fork, take, put, cps, call, select } from "redux-saga/effects";
+import { fork, take, put, select } from "redux-saga/effects";
 import gql from "graphql-tag";
 
 import { GraphQL } from "../../../graphql";
@@ -15,354 +12,19 @@ import modalActions from "../../modal";
 import types from "./../types";
 import actions from "../actions";
 
-import { CreditCardForm, AchForm } from "./paymentForm";
-import formatPersonDetails from "./formatPersonDetails";
+import RecoverSchedule from "../../../blocks/recover-schedules";
 
-import { order, schedule, charge } from "../../../methods/give/browser";
-import RecoverSchedules from "../../../blocks/recover-schedules";
+import validate from "./validate";
+import chargeTransaction from "./chargeTransaction";
+import submitPaymentDetails from "./submitPaymentDetails";
+import submitPersonDetails from "./submitPersonDetails";
 
-export const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
-
-// XXX break this file up into smaller files
-
-// at this point in time we have to do steps 1 - 3 of the
-// three step process to create a validation response
-// this validation process is required to ensure that the account
-// that is being used to make payments, is actually valid
-// see https://github.com/NewSpring/Apollos/issues/439 for discussion
-const ORDER_MUTATION = gql`
-  mutation order($data: String!, $id: ID, $instant: Boolean) {
-    response: createOrder(data: $data, id: $id, instant: $instant) {
-      url
-      error
-      success
-      code
-      transactionId
-    }
-  }
-`;
-
-export function* validate() {
-  const { give } = yield select(),
-    name = give.data.payment.name;
-
-  let success = true,
-    validationError = false,
-    transactionResponse;
-
-  // we strip all product and schedule data so the validation is
-  // just of the personal details + billing address
-  const modifiedGive = { ...give };
-  delete modifiedGive.transactions;
-  delete modifiedGive.schedules;
-
-
-  // step 1 (sumbit personal details)
-  // personal info is ready to be submitted
-  const formattedData = formatPersonDetails(modifiedGive);
-
-  // in order to make this a validation call, we need to set the amount
-  // to be 9
-  formattedData.amount = 0;
-
-  let error, url;
-  try {
-    // call the Meteor method to submit data to NMI
-    const { data: { response } } = yield GraphQL.mutate({
-      mutation: ORDER_MUTATION,
-      variables: {
-        data: JSON.stringify(formattedData),
-        id: null,
-        instant: false,
-      },
-    });
-    url = response.url;
-  } catch (e) { error = e; }
-
-  // step 2 (submit payment details)
-  yield submitPaymentDetails(modifiedGive.data, url);
-
-  if (url) {
-    // step 3 (trigger validation)
-    const token = url.split("/").pop();
-    try {
-      transactionResponse = yield cps(charge, token, name, null);
-    } catch (e) {
-      validationError = e;
-      success = false;
-    }
-  } else {
-    success = false;
-    validationError = error;
-  }
-
-  return {
-    success,
-    validationError,
-  };
-}
-
-// handle the transactions
-export function* chargeTransaction({ state }) {
-  if (state !== "submit") return;
-
-  let { give } = yield select(),
-    name = give.data.payment.name,
-    action = charge,
-    error = false,
-    id;
-
-  // set loading state
-  yield put(actions.loading());
-  // personal info is ready to be submitted
-  const formattedData = formatPersonDetails(give);
-
-  // if you have a saved account, NMI lets you "order" a schedule
-  // instead of order + charge
-  if (formattedData.savedAccount && Object.keys(give.schedules).length) {
-
-    // wrap the function for the same api
-    action = (token, name, id, callback) => {
-      GraphQL.mutate({
-        mutation: ORDER_MUTATION,
-        variables: {
-          data: JSON.stringify(formattedData),
-          instant: true,
-          id,
-        },
-      })
-        .then(x => callback(null, x))
-        .catch(callback);
-    };
-
-  } else {
-    let store = yield select();
-    give = store.give;
-
-    if (formattedData.savedAccount) {
-      // set people data and store transaction id
-      yield* submitPersonDetails(give, true);
-    }
-
-    store = yield select();
-    give = store.give;
-
-    // wait until we have the transaction url
-    if (!give.url) {
-      const { url } = yield take(types.SET_TRANSACTION_DETAILS);
-      give.url = url;
-    }
-  }
-
-  // get the token and name of the saved account
-  const token = give.url.split("/").pop();
-
-  if (Object.keys(give.schedules).length) {
-    // if there is not a saved account, charge the order
-    if (!formattedData.savedAccount) {
-      action = schedule;
-
-      if (give.data.payment.type === "cc") {
-        // saved accounts don't validate the payment by default
-        // so we make 3 blocking requests to validate the card :(
-        let { success, validationError } = yield* validate();
-
-        if (validationError) {
-          error = validationError;
-        }
-      }
-    }
-  }
-
-
-  if (give.scheduleToRecover && Object.keys(give.schedules).length) {
-    id = give.scheduleToRecover;
-  }
-
-  if (give.userId) id = give.userId; // used in native => web payments
-
-  let transactionResponse = {};
-  // submit transaction
-  try {
-    if (!error) {
-      transactionResponse = yield cps(action, token, name, id);
-    }
-  } catch (e) { error = e; }
-
-  // set error states
-  if (error) {
-    yield put(actions.error({ transaction: error }));
-
-    // remove loading state
-    yield put(actions.setState("error"));
-  } else {
-    // remove loading state
-    yield put(actions.setState("success"));
-
-    // if we activated an inactive schedule, remove it
-    if (give.scheduleToRecover && give.recoverableSchedules[give.scheduleToRecover]) {
-      yield put(actions.deleteSchedule(give.scheduleToRecover));
-      yield put(actions.deleteRecoverableSchedules(give.scheduleToRecover));
-    }
-
-    // if this was a named card (as in creating a saved account)
-    // lets force and update of the payment cards and set it in the store
-    // @TODO this is a race condition against updates in Rock
-    // we don't have a way to optimistcally update this without it being a
-    // hacky work around. I think this can wait until Apollo is closer
-    // to revist
-    if (name) {
-      const query = gql`
-        query GetSavedPayments {
-          savedPayments {
-            name
-            id
-            date
-            payment {
-              accountNumber
-              paymentType
-            }
-          }
-        }
-      `;
-
-      yield GraphQL.query({ query });
-    }
-  }
-}
-
-
-export function* submitPaymentDetails(data, url) {
-  /*
-
-    Oddity with NMI, when using a saved account for a subscription,
-    you only submit the order, not the charge, etc
-
-  */
-  if (data.payment.type === "cc" && !data.payment.cardNumber) return;
-  if (data.payment.type === "ach" && !data.payment.accountNumber) return;
-
-  const form = document.createElement("FORM");
-
-  let Component,
-    obj;
-
-  const { payment, personal } = data;
-
-  // format data and select component
-  if (payment.type === "ach") {
-    obj = {
-      name: `${personal.firstName} ${personal.lastName}`,
-      account: payment.accountNumber,
-      routing: payment.routingNumber,
-      type: payment.accountType,
-    };
-
-    Component = AchForm;
-  } else {
-    obj = {
-      number: payment.cardNumber,
-      exp: payment.expiration,
-      ccv: payment.ccv,
-    };
-
-    Component = CreditCardForm;
-  }
-
-  // create the fieldset
-  const FieldSet = React.createElement(Component, { ...obj });
-  // add fieldset to non rendered form
-  ReactDOM.render(FieldSet, form);
-
-  // @TODO test on older browsers
-  // store data in NMI's system
-  yield fetch(url, {
-    method: "POST",
-    body: new FormData(form),
-    mode: "no-cors",
-  })
-    .then((response) => {
-      // next()
-
-    })
-    .catch((e) => {
-      // @TODO error handling
-    });
-  yield delay(300); // ensure gift is in nmi's system beofre progressing
-  return;
-}
-
-export function* submitPersonDetails(give, autoSubmit) {
-  // personal info is ready to be submitted
-  const formattedData = formatPersonDetails(give);
-  /*
-
-    Oddity with NMI, when using a saved account for a subscription,
-    you only submit the order, not the charge, etc
-
-  */
-  if (formattedData.savedAccount && Object.keys(give.schedules).length) {
-    return;
-  }
-
-  let error, url;
-  try {
-    // call the Meteor method to submit data to NMI
-    const { data: { response } } = yield GraphQL.mutate({
-      mutation: ORDER_MUTATION,
-      variables: {
-        data: JSON.stringify(formattedData),
-        id: null,
-        instant: false,
-      },
-    });
-    if (response.error) error = response.error;
-    url = response.url;
-  } catch (e) { error = e; }
-
-  if (autoSubmit) {
-    /*
-
-      @NOTE
-        This is a hacky way to get around the submission required
-        by transnational gateway. We do a POST into the dark web of
-        NMI servers
-
-      @NOTE
-        This will always throw an error but we just catch it and ignore
-
-      @NOTE
-        DO NOT REMOVE THIS
-
-    */
-    // DONT REMOVE
-    // submit empty payment details to nmi
-    // fixes nasty IE / old browser / security bug
-    const form = document.createElement("FORM");
-    // create the fieldset
-    const FieldSet = React.createElement(CreditCardForm, {
-      number: "",
-      exp: "",
-    });
-    // add fieldset to non rendered form
-    ReactDOM.render(FieldSet, form);
-
-    const response = yield fetch(url, {
-      method: "POST",
-      body: new FormData(form),
-      mode: "no-cors",
-    })
-      .then(console.log("i made it"))
-      .catch((error) => {
-        // console.log("error dark", error);
-      });
-
-    yield delay(300); // ensure gift is in nmi's system beofre progressing
-  }
-
-  // update the store with the url
-  return yield put(actions.setDetails(url));
-}
+export {
+  validate,
+  chargeTransaction,
+  submitPaymentDetails,
+  submitPersonDetails,
+};
 
 // transaction processing flow controller
 function* createOrder() {
@@ -420,9 +82,7 @@ function* recoverTransactions() {
 
   let user = Meteor.userId();
 
-  if (!user) {
-    const { authorized } = yield take("ACCOUNTS.IS_AUTHORIZED");
-  }
+  if (!user) yield take("ACCOUNTS.IS_AUTHORIZED");
 
   user = Meteor.user();
 
@@ -455,19 +115,22 @@ function* recoverTransactions() {
   if (!schedules) schedules = [];
   hasRecovered = true;
   const bulkUpdate = {};
-  schedules = schedules.filter(x => !x.gateway);
+  schedules = schedules.filter((x) => !x.gateway);
   if (schedules.length) {
     for (const schedule of schedules) {
       // only recover schedules that are missing info (i.e. not turned off in Rock)
-      if (schedule.gateway) continue;
+      if (schedule.gateway) continue; // eslint-disable-line
 
       if (schedule.schedule.value === "Twice a Month") {
         schedule.schedule.value = null;
       }
-      bulkUpdate[schedule.id] = { ...{
-        start: Moment(schedule.start).format("YYYYMMDD"),
-        frequency: schedule.schedule.value,
-      }, ...schedule };
+      bulkUpdate[schedule.id] = {
+        ...{
+          start: moment(schedule.start).format("YYYYMMDD"),
+          frequency: schedule.schedule.value,
+        },
+        ...schedule,
+      };
     }
 
     let time = new Date();
@@ -488,7 +151,7 @@ function* recoverTransactions() {
       return;
     }
 
-    yield put(modalActions.render(RecoverSchedules));
+    yield put(modalActions.render(RecoverSchedule));
   }
 }
